requirements.txt
# Gas Town MEOW Stack - Molecule Marketplace Dependencies

# Core CLI framework
click>=8.0.0

# Database
sqlite3  # Built-in with Python

# Template processing
toml>=0.10.0
PyYAML>=6.0
Jinja2>=3.0.0

# HTTP client for API integrations
requests>=2.28.0

# JSON schema validation
jsonschema>=4.0.0

# File system operations
pathlib  # Built-in with Python

# Logging and utilities
python-dotenv>=0.19.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0